requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-snowflake
psycopg2-binary
snowflake-connector-python
apache-airflow